Creating chatbot documents page and related modals. Pneding edits on creating new document chunk
