+++ packages/app/components/ChatbotSettings/ChatbotDocuments.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import { Table, Radio, Button, Modal, Input, Form, message } from 'antd'
 import { useProfile } from '../../hooks/useProfile'
-import { CloseOutlined } from '@ant-design/icons'
+import EditDocumentModal from './EditChatbotDocumentModal'
 import axios from 'axios'
 
 interface ChatbotDocumentsProps {
@@ -55,26 +55,9 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
     },
     {
       title: 'Document Content',
-      dataIndex: ['metadata', 'original'],
-      key: 'pageContent',
-      width: 300,
-      render: (text) => (
-        <div
-          style={{
-            maxHeight: '150px',
-            overflow: 'auto',
-            whiteSpace: 'pre-wrap',
-          }}
-        >
-          {text}
-        </div>
-      ),
-    },
-    {
-      title: 'Edited Chunk',
       dataIndex: ['pageContent'],
-      key: 'editedChunk',
-      width: 300,
+      key: 'pageContent',
+      width: 500,
       render: (text) => (
         <div
           style={{
@@ -87,6 +70,23 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
         </div>
       ),
     },
+    // {
+    //   title: 'Edited Chunk',
+    //   dataIndex: ['pageContent'],
+    //   key: 'editedChunk',
+    //   width: 300,
+    //   render: (text) => (
+    //     <div
+    //       style={{
+    //         maxHeight: '150px',
+    //         overflow: 'auto',
+    //         whiteSpace: 'pre-wrap',
+    //       }}
+    //     >
+    //       {text}
+    //     </div>
+    //   ),
+    // },
     {
       title: 'Page Number',
       dataIndex: ['metadata', 'loc', 'pageNumber'],
@@ -154,10 +154,7 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
 
     const filtered = documents.filter((doc) => {
       const isNameMatch = doc.metadata.name.toLowerCase().includes(searchTerm)
-      const isUnsuccessful = doc.metadata.editedChunkText
-        .toLowerCase()
-        .includes('unsuccessful')
-      return isNameMatch || isUnsuccessful
+      return isNameMatch
     })
 
     setFilteredDocuments(filtered)
@@ -172,10 +169,7 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
     const searchTerm = search.toLowerCase()
     const filtered = updatedDocuments.filter((doc) => {
       const isNameMatch = doc.metadata.name.toLowerCase().includes(searchTerm)
-      const isUnsuccessful = doc.metadata.editedChunkText
-        .toLowerCase()
-        .includes('unsuccessful')
-      return isNameMatch || isUnsuccessful
+      return isNameMatch
     })
 
     setFilteredDocuments(filtered)
@@ -189,7 +183,7 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
             View Chatbot Documents
           </h3>
           <p className="text-[16px] font-medium text-gray-600">
-            View and manage the documents available to your chatbot
+            View and manage the document chunks from your documents
           </p>
         </div>
       </div>
@@ -197,11 +191,6 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
         <Form form={form} onFinish={addDocument}>
           <div className="flex items-center justify-between">
             <h2 className="text-2xl font-bold">New Document Chunk</h2>
-            <Button
-              type="text"
-              icon={<CloseOutlined />}
-              onClick={() => setAddModelOpen(false)}
-            />
           </div>
           <div className="mt-4">
             <Form.Item
@@ -225,9 +214,9 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
             >
               <Input.TextArea />
             </Form.Item>
-            <Form.Item label="Edited Chunk" name="editedChunk">
+            {/* <Form.Item label="Edited Chunk" name="editedChunk">
               <Input.TextArea />
-            </Form.Item>
+            </Form.Item> */}
             <Form.Item label="Source" name="source">
               <Input />
             </Form.Item>
@@ -267,115 +256,14 @@ export default function ChatbotDocuments({ courseId }: ChatbotDocumentsProps) {
       {editingRecord && (
         <EditDocumentModal
           editingRecord={editingRecord}
+          courseId={courseId}
+          chatbotToken={profile.chat_token.token}
           visible={editRecordModalVisible}
           setEditingRecord={setEditRecordModalVisible}
           onSuccessfulUpdate={updateDocumentInState}
         />
       )}
     </div>
-  )
-}
-
-/* eslint-disable react/prop-types */
-const EditDocumentModal = ({
-  editingRecord,
-  visible,
-  setEditingRecord,
-  onSuccessfulUpdate,
-}) => {
-  const [form] = Form.useForm()
-  const [selectedText, setSelectedText] = useState('content')
-
-  useEffect(() => {
-    if (editingRecord) {
-      form.setFieldsValue({
-        documentName: editingRecord.metadata.name,
-        content: editingRecord.metadata.original,
-        editedChunk: editingRecord.pageContent,
-        source: editingRecord.metadata.source,
-        pageNumber: editingRecord.metadata.loc.pageNumber,
-      })
-    }
-  }, [editingRecord, form])
-
-  const handleOk = async () => {
-    try {
-      const values = await form.validateFields()
-      const response = await fetch(
-        `/chat/${courseId}/${editingRecord.id}/documentChunk`,
-        {
-          method: 'PATCH',
-          headers: {
-            'Content-Type': 'application/json',
-            HMS_API_TOKEN: 'test_token',
-          },
-          body: JSON.stringify({
-            documentText:
-              selectedText === 'content' ? values.content : values.editedChunk,
-            metadata: {
-              name: values.documentName,
-              editedChunkText: values.editedChunk,
-              source: values.source,
-              loc: {
-                pageNumber: values.pageNumber,
-              },
-            },
-          }),
-        },
-      )
-
-      if (!response.ok) {
-        throw new Error('Network response was not ok')
-      }
-
-      const updatedDoc = await response.json()
-      onSuccessfulUpdate(updatedDoc)
-      setEditingRecord(false)
-      message.success('Document updated successfully.')
-    } catch (e) {
-      message.error('Failed to update document.')
-    }
-  }
-
-  const handleTextChange = (e) => {
-    setSelectedText(e.target.value)
-  }
-
-  return (
-    <Modal
-      title="Edit Document"
-      visible={visible}
-      onCancel={() => setEditingRecord(false)}
-      onOk={handleOk}
-      width={800}
-    >
-      <Form form={form} layout="vertical">
-        <Form.Item
-          label="Document Name"
-          name="documentName"
-          rules={[
-            { required: true, message: 'Please input the document name!' },
-          ]}
-        >
-          <Input />
-        </Form.Item>
-        <Form.Item label="Original Content" name="content">
-          <Input.TextArea style={{ height: 120 }} />
-        </Form.Item>
-        <Form.Item label="Edited chunk" name="editedChunk">
-          <Input.TextArea style={{ height: 120 }} />
-        </Form.Item>
-        <Form.Item label="Select Text to Embed">
-          <Radio.Group onChange={handleTextChange} value={selectedText}>
-            <Radio value="content">Original Content</Radio>
-            <Radio value="editedChunk">Edited Chunk</Radio>
-          </Radio.Group>
-        </Form.Item>
-        <Form.Item label="Source" name="source">
-          <Input />
-        </Form.Item>
-      </Form>
-    </Modal>
   )
 }
